Added crocodiles along the river and new crocodile's scene generation when starting a new manche. Crocodile can not respawn yet, feature cooming soon.

+++ versione_thread/settings.h
@@ -7,7 +7,7 @@ typedef struct gameRules
 {
     short time;
     bool enemies;
-    float BadCrocodile;
+    short BadCrocodile;
     short speed;
 } GameRules;
 #endif
